requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gTTS==2.2.2
h5py==2.10.0
idna==2.10
joblib==1.0.1
Keras-Preprocessing==1.1.2
Markdown==3.3.4
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
playsound==1.2.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/murta/Google%20Drive/Quick%20Access/Breeze/dependencies/whl/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
SpeechRecognition==3.8.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
